Skrur av mocking av infotrygd og kaller heller familie-mock-server.
