+++ src/test/kotlin/no/nav/familie/ba/sak/kjerne/verdikjedetester/FamilieBaSakKlient.kt
@@ -15,6 +15,7 @@ import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingKategori
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingType
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingUnderkategori
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
+import no.nav.familie.ba.sak.kjerne.dokument.domene.ManueltBrevRequest
 import no.nav.familie.ba.sak.kjerne.fagsak.FagsakRequest
 import no.nav.familie.ba.sak.kjerne.fagsak.RestBeslutningPåVedtak
 import no.nav.familie.ba.sak.kjerne.logg.Logg
@@ -159,5 +160,10 @@ class FamilieBaSakKlient(
         val uri = URI.create("$baSakUrl/api/e2e/task/$key/$value")
 
         return getForEntity(uri, headers)
+    }
+
+    fun forhaandsvisHenleggelseBrev(behandlingId: Long, manueltBrevRequest: ManueltBrevRequest): Ressurs<ByteArray>? {
+        val uri = URI.create("$baSakUrl/api/dokument/forhaandsvis-brev/${behandlingId}")
+        return postForEntity(uri, manueltBrevRequest, headers)
     }
 }
